Satellite image retrieval scripts.
Initial population: This is a script and a function that worked together to retrieve MODIS imagery from the NASA aqua and terra satellites. One image from each day was downloaded to a local directory. Both Aqua and Terra were looked at. Whichever one was A) below <50 cloud fraction and B) if both had less than <50 cloud faction, the lowest cloud fraction was chosen.